move security data in a proper package
